Configuração do Envio de email de cadastro

+++ database/migrations/2022_10_05_193836_create_events_table.php
@@ -15,7 +15,7 @@ class CreateEventsTable extends Migration
     {
         Schema::create('events', function (Blueprint $table) {
             $table->string('placa');
-            $table->integer('renavam');
+            $table->bigInteger('renavam');
             $table->text('modelo');
             $table->text('marca');
             $table->year('ano');
